+++ src/main/java/com/kr1s1s/subtlyd/data/ModelProviderSD.java
@@ -9,6 +9,7 @@ import net.minecraft.client.data.models.ItemModelGenerators;
 import net.minecraft.client.data.models.MultiVariant;
 import net.minecraft.client.data.models.model.ModelTemplates;
 import net.minecraft.client.data.models.model.TextureMapping;
+import net.minecraft.client.data.models.model.TexturedModel;
 import net.minecraft.data.BlockFamily;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.Item;
@@ -23,9 +24,16 @@ public class ModelProviderSD extends FabricModelProvider {
 
     @Override
     public void generateBlockStateModels(BlockModelGenerators blockModelGenerator) {
-        blockModelGenerator.family(BlocksSD.SNOW_BRICKS).generateFor(new BlockFamily.Builder(BlocksSD.SNOW_BRICKS).stairs(BlocksSD.SNOW_BRICK_STAIRS).slab(BlocksSD.SNOW_BRICK_SLAB).getFamily());
+        blockModelGenerator.family(BlocksSD.SNOW_BRICKS).generateFor(new BlockFamily.Builder(BlocksSD.SNOW_BRICKS).stairs(BlocksSD.SNOW_BRICK_STAIRS).slab(BlocksSD.SNOW_BRICK_SLAB).wall(BlocksSD.SNOW_BRICK_WALL).getFamily());
+        blockModelGenerator.family(BlocksSD.POLISHED_STONE).generateFor(new BlockFamily.Builder(BlocksSD.POLISHED_STONE).stairs(BlocksSD.POLISHED_STONE_STAIRS).slab(BlocksSD.POLISHED_STONE_SLAB).wall(BlocksSD.POLISHED_STONE_WALL).getFamily());
+        blockModelGenerator.family(BlocksSD.POLISHED_DRIPSTONE).generateFor(new BlockFamily.Builder(BlocksSD.POLISHED_DRIPSTONE).stairs(BlocksSD.POLISHED_DRIPSTONE_STAIRS).slab(BlocksSD.POLISHED_DRIPSTONE_SLAB).wall(BlocksSD.POLISHED_DRIPSTONE_WALL).getFamily());
+        blockModelGenerator.family(BlocksSD.STONE_TILES).generateFor(new BlockFamily.Builder(BlocksSD.STONE_TILES).stairs(BlocksSD.STONE_TILE_STAIRS).slab(BlocksSD.STONE_TILE_SLAB).wall(BlocksSD.STONE_TILE_WALL).getFamily());
+
         blockModelGenerator.createTrivialCube(BlocksSD.CHARCOAL_BLOCK);
         createCubeFromVanilla(Blocks.IRON_BARS, BlocksSD.IRON_GRATE, blockModelGenerator);
+        blockModelGenerator.createAxisAlignedPillarBlock(BlocksSD.STONE_PILLAR, TexturedModel.COLUMN);
+        blockModelGenerator.createTrivialCube(BlocksSD.CHISELED_STONE);
+        blockModelGenerator.createTrivialCube(BlocksSD.CHISELED_DRIPSTONE);
     }
 
     @Override